+++ system/logic/desktop.js
@@ -1,0 +1,6 @@
+$(function () {
+
+  // Wallpaper
+  $(".system__desktop").backstretch("shell/wallpaper/tokyoSkytree-observationDeck/14.darwinfish105.jpg");
+
+});