+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+0.1.2 / 03.18.2015
+==================
+* updated dependecies, folder structure, and initiated a major rewrite
+* added clock, online status, avatar, and "Start" button to taskbar
+* implemented nice Start menu function
+* `.system__base` changes look based on hour of day
+
+
+
 0.1.0 / 03.15.2015
 ==================
 * created a CHANGELOG